api/requirements.txt
flask
flask-cors
pytest
coverage
waitress
psycopg2-binary